deepen: merge pill-pct and materiality into one financial-render-helpers module co-located with the templates
Replaces two shallow lib/email/ modules (pill-pct.ts + materiality.ts,
244 LOC combined) with a single deep module co-located with the two
minimalist templates that consume them
(components/ui/email/templates/financial-render-helpers.ts).

Both former modules had the SAME two callers — 10k-minimalist-template.tsx
and 10q-minimalist-template.tsx. Two adapters (files) crossing the same
seam to serve the same callers = one hypothetical seam, not two real
ones. Each caller had to reach across the lib/email boundary for a pair
of render helpers that only ever rendered inside these two components.

Category: in-process (pure computation over HTML strings and structured
JSON, no I/O). No adapter needed. The deepened module's interface is
tested through the templates that render its output.

Interface (5 exports, one file):
- wrapPercentsInPills(html): post-process markdown-rendered HTML so
  "+12.3%" / "-3.9pp" tokens become colored pill spans
- extractMaterialitySignal(summaryJSON): pull the optional
  materialitySignal field out of Grok's structured output with a safe
  noise default
- materialityToBadge(signal): map score -> label + colorKey, null on
  noise (caller falls back to form-type label)
- buildMaterialityFeedbackMailto({ticker, formType, ...}): build the
  "Wrong call?" mailto: url that anchors the feedback loop
- MaterialityScore / MaterialitySignal / MaterialityBadge / MaterialityBadgeColorKey types

Locality: 10-K and 10-Q templates now import from ./financial-render-helpers
(one relative-path import instead of two cross-lib absolute-path imports).
A future rendering tweak reads and edits one file instead of jumping
between lib/email/pill-pct.ts, lib/email/materiality.ts, and the templates.

Leverage: the templates' import surface for render helpers shrinks from
two modules to one; the type contract for materiality (MaterialitySignal
+ MaterialityBadge) lives next to the function that produces the badge
from the signal.

Dead-code cleanup: the former pill-pct.ts also exported escapeHtmlMinimal
with zero callers anywhere in the codebase. Not carried forward.

Tests deleted:
- __tests__/email/materiality.test.ts (142 LOC): asserted on the return
  shape of extractMaterialitySignal / materialityToBadge / buildMaterialityFeedbackMailto
  directly — past the JSX interface. Same shape as the ADR-0002 /
  Why-It-Matters-Guard precedent (extracted-for-testability pure functions
  whose real caller is the component that renders their output).

Tests retained (unchanged), covers every branch through the JSX interface:
- __tests__/email/materiality-badge-rendering.test.tsx: end-to-end
  rendering assertions on the 10-K and 10-Q templates for high / medium /
  low / noise / absent / invalid materiality signals; also asserts the
  mailto: URL encoding and the form-type fallback ("ANNUAL REPORT" /
  "QUARTERLY REPORT"). Every case the deleted unit test covered now
  crosses the same seam that production renders through.

pill-pct.ts had no dedicated test file — its wrapPercentsInPills export
was covered only by the templates that call it, and their existing test
coverage (10q-pill-delta.test.tsx et al.) continues to exercise it.

All 112 tests in materiality-* / pill-* / 10k-minimalist / 10q-minimalist
patterns pass on this branch. The 8 pre-existing email-test-suite
failures (email-notifications, duplicate-detection, form144-data-extractor,
quarterly-earnings-email, sections.test.tsx, email-validation-security,
db-email-integration, lib/email/__tests__/templates.test.ts) also fail
on main — env / react-email VM-module / puppeteer setup gaps, not
touched by this diff.
